Add command to 'image disk download'
