--- bin/disk/download/examples.md
@@ -1,3 +1,0 @@
-```bash
-{{command_name}} --disk db-disk --destination-file ./my-disk.vhdx
-```
