+++ bin/disk/index.js
@@ -8,13 +8,12 @@ const resource = {
     defaultQuery: '[].{id:_id,name:name,type:type,size:size,state:state,tags:join(\',\',keys(tag || `{}`) ) }',
     url: () => 'disk',
     plugins: genericDefaults.plugins,
-    extraCommands: ['resize', 'resume', 'transfer'],
+    extraCommands: ['resize', 'resume', 'transfer', 'download'],
     title: 'disk',
 };
 
 const category = genericResource(resource);
 
 category.addChild(require('./create')(resource));
-category.addChild(require('./download')(resource));
 
 module.exports = category;
