+++ bin/generic/download/examples.md
@@ -1,0 +1,3 @@
+```bash
+{{command_name}} --disk db-disk --destination-file ./my-disk.vhdx {{downloadParams | ''}}
+```
