--- bin/image/disk/examples.md
@@ -1,3 +1,0 @@
-```bash
-{{command_name}} --image dev-image
-```
