+++ docs/disk.md
@@ -7,12 +7,12 @@
   * [h1 disk delete](#h1-disk-delete) - Delete disk
   * [h1 disk history](#h1-disk-history) - History of disk
   * [h1 disk rename](#h1-disk-rename) - Rename disk
-  * [h1 disk create](#h1-disk-create) - Create disk
+  * [h1 disk download](#h1-disk-download) - Download disk to a .vhdx file
   * [h1 disk show](#h1-disk-show) - Show disk
   * [h1 disk resize](#h1-disk-resize) - Resize disk
   * [h1 disk resume](#h1-disk-resume) - Resume create upload of disk
   * [h1 disk transfer](#h1-disk-transfer) - Transfer disk to other project
-  * [h1 disk download](#h1-disk-download) - Download disk to a .vhdx file
+  * [h1 disk create](#h1-disk-create) - Create disk
   * [h1 disk tag](#h1-disk-tag) - Manage your tag
     * [h1 disk tag list](#h1-disk-tag-list) - List tag
     * [h1 disk tag add](#h1-disk-tag-add) - Add a tag to disk
@@ -139,42 +139,31 @@ h1 disk rename --disk my-disk --new-name my-renamed-disk
 | ```--disk DISK``` |  | Disk ID or name |
 | ```--new-name NEW-NAME``` |  | New name |
 
-## h1 disk create
+## h1 disk download
 
-Create disk
+Download disk to a .vhdx file
 
 ### Syntax
 
-```h1 disk create | --name NAME --type TYPE [--size SIZE] [--source-file SOURCE-FILE] [--no-progress] [--tag TAG [--tag TAG ...]]```
-### Examples
-
-#### Create 10 GB SSD disk
+```h1 disk download | --disk DISK --destination-file DESTINATION-FILE [--no-progress]```
+### Example
 
 ```bash
-h1 disk create --name db-disk --type ssd --size 10
-```
-
-#### Create disk from .vhdx file
-
-```bash
-h1 disk create --name new-disk --size 1 --type ssd --source-file ./my-disk.vhdx
+h1 disk download --disk db-disk --destination-file ./my-disk.vhdx
 ```
 
 ### Required arguments
 
 | Name | Default | Description |
 | ---- | ------- | ----------- |
-| ```--name NAME``` |  | Disk name |
-| ```--type TYPE``` |  | Disk type ID or name |
+| ```--disk DISK``` |  | Disk ID or name |
+| ```--destination-file DESTINATION-FILE``` |  | Path to .vhdx file to save |
 
 ### Optional arguments
 
 | Name | Default | Description |
 | ---- | ------- | ----------- |
-| ```--size SIZE``` |  | Disk size in GiB. Required if no source file is specified |
-| ```--source-file SOURCE-FILE``` |  | Path to .vhdx file to import |
 | ```--no-progress``` |  | Disable progress bar |
-| ```--tag TAG [--tag TAG ...]``` |  | Key=value of tag. The parameter may occur repeatedly |
 
 ## h1 disk show
 
@@ -255,31 +244,42 @@ h1 disk transfer --disk test-disk --new-project OtherProject
 | ```--disk DISK``` |  | Disk ID or name |
 | ```--new-project NEW-PROJECT``` |  | New name |
 
-## h1 disk download
+## h1 disk create
 
-Download disk to a .vhdx file
+Create disk
 
 ### Syntax
 
-```h1 disk download | --disk DISK --destination-file DESTINATION-FILE [--no-progress]```
-### Example
+```h1 disk create | --name NAME --type TYPE [--size SIZE] [--source-file SOURCE-FILE] [--no-progress] [--tag TAG [--tag TAG ...]]```
+### Examples
+
+#### Create 10 GB SSD disk
 
 ```bash
-h1 disk download --disk db-disk --destination-file ./my-disk.vhdx
+h1 disk create --name db-disk --type ssd --size 10
+```
+
+#### Create disk from .vhdx file
+
+```bash
+h1 disk create --name new-disk --size 1 --type ssd --source-file ./my-disk.vhdx
 ```
 
 ### Required arguments
 
 | Name | Default | Description |
 | ---- | ------- | ----------- |
-| ```--disk DISK``` |  | Disk name or ID |
-| ```--destination-file DESTINATION-FILE``` |  | Path to .vhdx file to save |
+| ```--name NAME``` |  | Disk name |
+| ```--type TYPE``` |  | Disk type ID or name |
 
 ### Optional arguments
 
 | Name | Default | Description |
 | ---- | ------- | ----------- |
+| ```--size SIZE``` |  | Disk size in GiB. Required if no source file is specified |
+| ```--source-file SOURCE-FILE``` |  | Path to .vhdx file to import |
 | ```--no-progress``` |  | Disable progress bar |
+| ```--tag TAG [--tag TAG ...]``` |  | Key=value of tag. The parameter may occur repeatedly |
 
 ## h1 disk tag
